+++ tram/cluster/__init__.py
@@ -1,0 +1,4 @@
+from tram.cluster.coordinator import ClusterCoordinator
+from tram.cluster.registry import NodeRegistry
+
+__all__ = ["ClusterCoordinator", "NodeRegistry"]